backend/pyproject.toml
[project]
name = "stocks_analyzer_be_01"
version = "0.1.0"
description = "stocks_analyzer_be_01 using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.7.2",
    "ddgs>=9.10.0",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "langchain-groq>=0.3.8",
    "langchain-openai>=0.3.23",
    "pydantic>=2.11.10",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-json-logger>=4.0.0",
    "redis>=7.1.0",
    "yfinance>=1.1.0",
]

[project.scripts]
stocks_analyzer_be_01 = "app.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.crewai]
type = "crew"

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
